Basic file transfer over UDP almost working
It's not reliable transfer yet, so no inorder delivery, ack, nor sliding
window. So long way to go, but the basic transfer process is nearly
there. It can send a request, look to see if the file exists on the
server, and send back if it is. I can't get the writing part done
properly yet though. It just writes the output of bytes.toString as the
file name, with no contents in the actual file.