Basic File Transfer Working
Some minor derpiness on writing the contents, just to add in come code
to parse the packet to get at the packet, then use fileinputstream to
write to the file. Need some time on wireshark to see how to get at it,
but I don't think that will be too hard to implement.